Add somewhat useful error spans
We don't have source spans but we can identify the part of the mnemonic
or argument in which an error occurs. Assuming this to be a common
encoding of opcodes (which of course it is) a caller can map them to
other useful spans.

Where strings are presented to the actual compiler we could also compute
better spans within the arguments but that is not part of an MVP
solution.